Refactor settings
Remove sourcemap generation by default
