+++ rollup.config.js
@@ -1,28 +1,28 @@
 import typescript from '@rollup/plugin-typescript';
 import resolve from '@rollup/plugin-node-resolve';
 import commonjs from '@rollup/plugin-commonjs';
-import json from '@rollup/plugin-json';
 import svelte from "rollup-plugin-svelte";
 import autoPreprocess from "svelte-preprocess";
 import copy from "rollup-plugin-copy";
+import { env } from "process";
 
 export default {
   input: 'src/index.ts',
   output: {
     file: 'dist/main.js',
-    sourcemap: 'inline',
     format: 'cjs',
     exports: 'default'
   },
-  external: ['obsidian'],
+  external: ['obsidian', "path"],
   plugins: [
-    json(),
-    typescript(),
+    typescript({ sourceMap: env.env === "DEV" }),
     resolve({
       browser: true,
       dedupe: ["svelte"],
     }),
-    commonjs(),
+    commonjs({
+      include: "node_modules/**"
+    }),
     svelte({
       preprocess: autoPreprocess(),
     }),
